rewrite TestModifyColumn
